refactor: reduced noise in events' show
