+++ event-log/src/main/scala/io/renku/eventlog/events/categories/globalcommitsyncrequest/EventHandler.scala
@@ -15,6 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package io.renku.eventlog.events.categories.globalcommitsyncrequest
 
 import cats.Show
